requirements.txt
python-telegram-bot==20.3
openai==0.27.4            
requests==2.31.0
beautifulsoup4==4.12.2